+++ .github/workflows/TS_Delivery_dev.yml
@@ -2,6 +2,8 @@ name: TS_Delivery_dev
 description: "should verify message loss when receiving via 200 streams"
 
 on:
+  pull_request:
+    branches: [deploy]
   schedule:
     - cron: "15,45 * * * *" # Runs at 15 and 45 minutes past each hour
   workflow_dispatch:
