+++ .github/workflows/TS_Gm_dev.yml
@@ -2,6 +2,8 @@ name: TS_Gm_dev
 description: "should verify performance of the library in the dev network"
 
 on:
+  pull_request:
+    branches: [deploy]
   schedule:
     - cron: "15,45 * * * *" # Runs at 15 and 45 minutes past each hour
   workflow_dispatch:
