+++ .github/workflows/package-manager-test.yml
@@ -1,10 +1,8 @@
 name: Package Manager Compatibility Test
 
 on:
-  push:
-    branches: [dev]
   pull_request:
-    branches: [dev]
+    branches: [deploy]
 
 jobs:
   test-package-managers:
