+++ bots/gm-bot/package.json
@@ -1,12 +1,12 @@
 {
   "name": "gm-bot",
+  "version": "0.0.5",
   "private": true,
   "type": "module",
   "scripts": {
     "build": "tsc",
-    "dev": "tsc -w & sleep 1 && NODE_NO_WARNINGS=1 node --watch dist/index.js",
-    "start": "node dist/index.js",
-    "test": "vitest"
+    "check": "tsx src/check.ts",
+    "start": "tsx src/index.ts"
   },
   "dependencies": {
     "@xmtp/node-sdk": "1.0.0-rc2",
@@ -17,8 +17,7 @@
   "devDependencies": {
     "@types/node": "^20.14.2",
     "dotenv": "^16.4.5",
-    "typescript": "^5.4.5",
-    "vitest": "^3.0.6"
+    "typescript": "^5.4.5"
   },
   "engines": {
     "node": ">=20"
