+++ bots/gm-bot/src/index.ts
@@ -24,11 +24,8 @@ async function main() {
   console.log(`Creating client on the '${env}' network...`);
   let volumePath = process.env.RAILWAY_VOLUME_MOUNT_PATH ?? ".data/xmtp";
   // Ensure the volume path directory exists
-  try {
+  if (await fs.stat(volumePath).catch(() => false)) {
     await fs.mkdir(volumePath, { recursive: true });
-    console.log(`Ensured directory exists: ${volumePath}`);
-  } catch (error) {
-    console.error(`Error creating directory ${volumePath}:`, error);
   }
   if (process.env.RAILWAY_VOLUME_MOUNT_PATH) {
     console.log(
