+++ package.json
@@ -3,9 +3,6 @@
   "version": "0.0.5",
   "private": true,
   "type": "module",
-  "workspaces": [
-    "bots/*"
-  ],
   "scripts": {
     "bot:gm": "tsx  bots/gm/index.ts",
     "bot:test": "tsx  bots/test/index.ts",
@@ -66,7 +63,7 @@
     "tsconfig-paths": "^4.2.0",
     "typescript": "^5.7.3",
     "typescript-eslint": "^8.22.0",
-    "vitest": "^3.0.6"
+    "vitest": "^3.0.8"
   },
   "packageManager": "yarn@4.6.0",
   "engines": {
